Added viewmodels package
