+++ app/src/main/java/app/milanherke/mystudiez/AddEditTaskActivity.kt
@@ -24,6 +24,7 @@ import app.milanherke.mystudiez.fragments.TasksFragment
 import app.milanherke.mystudiez.fragments.UnsavedChangesDialogFragment
 import app.milanherke.mystudiez.models.Subject
 import app.milanherke.mystudiez.models.Task
+import app.milanherke.mystudiez.viewmodels.activities.AddEditTaskViewModel
 import com.google.firebase.database.DatabaseError
 import kotlinx.android.synthetic.main.activity_add_edit_task.*
 import kotlinx.android.synthetic.main.content_add_edit_task.*
