+++ app/src/main/java/app/milanherke/mystudiez/fragments/ExamDetailsFragment.kt
@@ -13,6 +13,7 @@ import app.milanherke.mystudiez.*
 import app.milanherke.mystudiez.FragmentBackStack
 import app.milanherke.mystudiez.models.Exam
 import app.milanherke.mystudiez.models.Subject
+import app.milanherke.mystudiez.viewmodels.fragments.ExamDetailsViewModel
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.fragment_exam_details.*
