+++ app/src/main/java/app/milanherke/mystudiez/viewmodels/fragments/ExamDetailsViewModel.kt
@@ -1,7 +1,8 @@
-package app.milanherke.mystudiez
+package app.milanherke.mystudiez.viewmodels.fragments
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
+import app.milanherke.mystudiez.FirebaseUtils
 import app.milanherke.mystudiez.models.Exam
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
