+++ app/src/main/java/app/milanherke/mystudiez/viewmodels/fragments/ExamsViewModel.kt
@@ -1,10 +1,11 @@
-package app.milanherke.mystudiez
+package app.milanherke.mystudiez.viewmodels.fragments
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import app.milanherke.mystudiez.models.Task
+import app.milanherke.mystudiez.FirebaseUtils
+import app.milanherke.mystudiez.models.Exam
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ValueEventListener
@@ -16,14 +17,14 @@ import kotlinx.coroutines.launch
 
 /**
  * A simple [AndroidViewModel] subclass.
- * This ViewModel was created to get all [Task] objects from the database
- * and belongs to [TasksViewModel].
+ * This ViewModel was created to get all [Exam] objects from the database
+ * and belongs to [app.milanherke.mystudiez.fragments.ExamsFragment].
  */
-class TasksViewModel(application: Application) : AndroidViewModel(application) {
+class ExamsViewModel(application: Application) : AndroidViewModel(application) {
 
-    private val tasksList = MutableLiveData<ArrayList<Task>>()
-    val tasksListLiveData: LiveData<ArrayList<Task>>
-        get() = tasksList
+    private val examsList = MutableLiveData<ArrayList<Exam>>()
+    val examsListLiveData: LiveData<ArrayList<Exam>>
+        get() = examsList
 
     /**
      * Classes that use this ViewModel's functions
@@ -37,17 +38,17 @@ class TasksViewModel(application: Application) : AndroidViewModel(application) {
     }
 
     /**
-     * Gets all tasks from the database
-     * and passes the result to [tasksList]
+     * Gets all exams from the database
+     * and passes the result to [examsList].
      *
      * @param listener [DataFetching] interface to handle interaction events
      */
-    fun loadTasks(listener: DataFetching) {
+    fun loadExams(listener: DataFetching) {
         GlobalScope.launch {
             listener.onLoad()
-            val tasks: ArrayList<Task> = arrayListOf()
+            val exams: ArrayList<Exam> = arrayListOf()
             val database = Firebase.database
-            val ref = database.getReference("tasks/${FirebaseUtils.getUserId()}")
+            val ref = database.getReference("exams/${FirebaseUtils.getUserId()}")
             ref.addListenerForSingleValueEvent(object : ValueEventListener {
                 override fun onCancelled(e: DatabaseError) {
                     listener.onFailure(e)
@@ -55,14 +56,14 @@ class TasksViewModel(application: Application) : AndroidViewModel(application) {
 
                 override fun onDataChange(dataSnapshot: DataSnapshot) {
                     for (subjectSnapshot in dataSnapshot.children) {
-                        for (taskSnapshot in subjectSnapshot.children) {
-                            val task = taskSnapshot.getValue<Task>()
-                            if (task != null) {
-                                tasks.add(task)
+                        for (examSnapshot in subjectSnapshot.children) {
+                            val exam = examSnapshot.getValue<Exam>()
+                            if (exam != null) {
+                                exams.add(exam)
                             }
                         }
                     }
-                    tasksList.postValue(tasks)
+                    examsList.postValue(exams)
                     listener.onSuccess()
                 }
             })
