+++ app/src/main/java/app/milanherke/mystudiez/viewmodels/fragments/SubjectsViewModel.kt
@@ -1,4 +1,4 @@
-package app.milanherke.mystudiez
+package app.milanherke.mystudiez.viewmodels.fragments
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
